+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 3.11.0
 * Fixed closing proxy clients
-* Added `ydb.Connection.WithCustomOptions()`
+* Added `ydb.Connection.With(opts ...ydb.CustomOption)` for taking proxy `ydb.Connection` with some redefined options
 * Added `ydb.MetaRequestType` and `ydb.MetaTraceID` aliases to internal `meta` package constants
 
 ## 3.10.0